feat: Updated styling alignment. chore: Renamed Folders to camel case and files to pascal case. chore: Created a widgets file for th ControlPanel
