+++ src/layout/fullPage/Layout.jsx
@@ -1,10 +1,10 @@
 // src/layout/fullPage/Layout.js
 import React, { useContext } from "react";
 import { Container, Row, Col } from "react-bootstrap";
-import Routing from "../../Router/Routing";
+import Routing from "../../router/Routing";
 import SideBar from "../default/SideBar";
 import Header from "../default/Header";
-import { DisplaySidebarContext } from "../../Components/ContextHook/useDisplayContext";
+import { DisplaySidebarContext } from "../../Components/contextHook/useDisplayContext";
 
 function Layout() {
   const { displaySidebarPanel } = useContext(DisplaySidebarContext);
