+++ src/pages/AppraisalActivity/ActivityTableData/TableHandler.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import TableComponent from './TableComponent';
-import AddActivityForm from '../AddActivity/ActivityForm';
+import AddActivityForm from '../addActivity/ActivityForm';
 import fetchData from '../../../services/DataService';
 import { useNavigate } from 'react-router-dom';
 import { Button } from 'react-bootstrap';
